OPIK-610 move client creation to provider class
